refactor(config): use POSTURE_MATRIX as single source of truth for send_email gate
Replace hardcoded posture match in validate_smtp_required with
base_allows() from rimap-core, eliminating drift risk between
config validation and the posture matrix.

Move POSTURE_MATRIX and base_allows() into rimap-core so that both
rimap-config and rimap-authz can query posture truth without a circular
dependency. rimap-authz delegates its own base_allows() to the core
implementation.
